Tease apart speech lib
